Speeding up map load. Not tested
