+++ PersistentEmpiresLib/PersistentEmpiresLib/PersistentEmpiresMission/MissionBehaviors/StockpileMarketComponent.cs
@@ -6,11 +6,85 @@ using PersistentEmpiresLib.SceneScripts;
 using System;
 using System.Collections.Generic;
 using System.Linq;
+using System.Xml;
+using TaleWorlds.Core;
 using TaleWorlds.Library;
+using TaleWorlds.ModuleManager;
 using TaleWorlds.MountAndBlade;
+using TaleWorlds.ObjectSystem;
 
 namespace PersistentEmpiresLib.PersistentEmpiresMission.MissionBehaviors
 {
+    public class MarketItem
+    {
+        public ItemObject Item;
+        public int Stock;
+        public int MaximumPrice;
+        public float CurrentPrice;
+        public int MinimumPrice;
+        public int Constant;
+        public int Stability;
+        public int Tier;
+        // X is stock Y is price x^(1/stability) * y = k
+        public MarketItem(string itemId, int maximumPrice, int minimumPrice, int stability, int tier)
+        {
+            this.Item = MBObjectManager.Instance.GetObject<ItemObject>(itemId);
+            this.MaximumPrice = maximumPrice;
+            this.MinimumPrice = minimumPrice;
+            this.Constant = MaximumPrice;
+            this.Stability = stability;
+            this.CurrentPrice = MathF.Pow(this.Constant, 1f / this.Stability);
+            this.Tier = tier;
+        }
+        public void UpdateReserve(int newStock)
+        {
+            if (newStock > 999) newStock = 999;
+            if (newStock < 0) newStock = 0;
+
+            if (newStock < 1)
+            {
+                this.CurrentPrice = MathF.Pow(this.Constant, 1f / this.Stability);
+            }
+            else
+            {
+                this.CurrentPrice = MathF.Pow(this.Constant / (float)newStock, 1f / this.Stability);
+            }
+            this.Stock = newStock;
+        }
+        public int BuyPrice()
+        {
+            float fakeStock = this.Stock;
+            if (this.Stock < 2) return this.MaximumPrice;
+            float denominator = MathF.Pow(fakeStock - 1, 1f / this.Stability);
+            float numerator = this.Constant;
+            int buyPrice = Math.Abs((int)((numerator / denominator) - MathF.Pow(this.CurrentPrice, 1f / this.Stability)));
+            if (buyPrice < this.MinimumPrice) buyPrice = this.MinimumPrice;
+            return buyPrice;
+        }
+        public int SellPrice()
+        {
+            float fakeStock = this.Stock;
+            if (this.Stock < 1) fakeStock = 1;
+            float denominator = MathF.Pow(fakeStock + 1, 1f / this.Stability);
+            float numerator = this.Constant;
+            int price = Math.Abs((int)(MathF.Pow(this.CurrentPrice, 1f / this.Stability) - (numerator / denominator)));
+            if (price < this.MinimumPrice) price = (this.MinimumPrice * 90) / 100;
+            return (price * 85) / 100;
+        }
+    }
+    public class CraftingBox
+    {
+        public ItemObject BoxItem;
+        public int MinTierLevel;
+        public int MaxTierLevel;
+        public CraftingBox(string itemId, string minTierLevel, string maxTierLevel)
+        {
+            this.BoxItem = MBObjectManager.Instance.GetObject<ItemObject>(itemId);
+            this.MinTierLevel = int.Parse(minTierLevel);
+            this.MaxTierLevel = int.Parse(maxTierLevel);
+        }
+    }
+
     public class StockpileMarketComponent : MissionNetwork
     {
         public long LastSaveAt = DateTimeOffset.Now.ToUnixTimeSeconds();
@@ -26,12 +100,71 @@ namespace PersistentEmpiresLib.PersistentEmpiresMission.MissionBehaviors
         public event StockpileMarketUpdateMultiHandler OnStockpileMarketUpdateMultiHandler;
 
         Dictionary<MissionObject, List<NetworkCommunicator>> openedInventories;
+
+        public static List<MarketItem> MarketItems { get; set; } = new List<MarketItem>();
+        public static List<CraftingBox> CraftingBoxes { get; set; } = new List<CraftingBox>();
+        public static string XmlFile = "examplemarket"; // itemId*minimum*maximum,itemId*minimum*maximum
+        private static string ModuleFolder = "PersistentEmpires";
+
         public override void OnBehaviorInitialize()
         {
             base.OnBehaviorInitialize();
             this.AddRemoveMessageHandlers(GameNetwork.NetworkMessageHandlerRegisterer.RegisterMode.Add);
             openedInventories = new Dictionary<MissionObject, List<NetworkCommunicator>>();
+#if SERVER
+            LoadXmlConfig();
+#endif
         }
+
+#if SERVER
+        private void LoadXmlConfig()
+        {
+            string xmlPath = ModuleHelper.GetXmlPath(ModuleFolder, "Markets/" + XmlFile);
+            XmlDocument xmlDocument = new XmlDocument();
+            xmlDocument.Load(xmlPath);
+
+            LoadMarketItems(xmlDocument.SelectSingleNode("/Market/Tier1Items").InnerText, 1);
+            LoadMarketItems(xmlDocument.SelectSingleNode("/Market/Tier2Items").InnerText, 2);
+            LoadMarketItems(xmlDocument.SelectSingleNode("/Market/Tier3Items").InnerText, 3);
+            LoadMarketItems(xmlDocument.SelectSingleNode("/Market/Tier4Items").InnerText, 4);
+            LoadCraftingBoxes(xmlDocument.SelectSingleNode("/Market/CraftingBoxes").InnerText);
+        }
+
+        protected void LoadMarketItems(string innerText, int tier)
+        {
+
+            if (innerText == "") return;
+            foreach (string marketItemStr in innerText.Trim().Split('|'))
+            {
+                string[] values = marketItemStr.Split('*');
+                string itemId = values[0];
+                int minPrice = int.Parse(values[1]);
+                int maxPrice = int.Parse(values[2]);
+                int stability = values.Length > 4 ? int.Parse(values[3]) : 10;
+
+                ItemObject item = MBObjectManager.Instance.GetObject<ItemObject>(itemId);
+                if (item == null)
+                {
+                    Debug.Print(" ERROR IN MARKET SERIALIZATION " + XmlFile + " ITEM ID " + itemId + " NOT FOUND !!! ", 0, Debug.DebugColor.Red);
+                }
+                MarketItems.Add(new MarketItem(itemId, maxPrice, minPrice, stability, tier));
+            }
+        }
+
+        protected void LoadCraftingBoxes(string innerText)
+        {
+            if (innerText == "") return;
+            foreach (string craftingBoxStr in innerText.Trim().Split('|'))
+            {
+                string[] values = craftingBoxStr.Split('*');
+                string itemId = values[0];
+                string minTierLevel = values[1];
+                string maxTierLevel = values[2];
+                CraftingBoxes.Add(new CraftingBox(itemId, minTierLevel, maxTierLevel));
+            }
+        }
+#endif
+
         public override void OnRemoveBehavior()
         {
             base.OnRemoveBehavior();
